scss-스타일링2
(1) footer 추가
