+++ src/components/Layout.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from "react"
+import MenuList from "./Menu";
 
 interface LayoutProps {
   children: ReactNode;
@@ -6,6 +7,9 @@ interface LayoutProps {
 
 export const Layout = ({children}: LayoutProps) => {
   return (
-    <div className="layout">{children}</div>
+    <div className="layout">
+      {children}
+      <MenuList />
+    </div>
   )
 }
