+++ src/index.scss
@@ -1,18 +1,17 @@
 $background: #1e293b;
 $activeBg: #3c495f5f;
-$white: #fff;
+$white: #ffffff;
 $border: #708090;
 $primaryColor: #60a5fa;
 $activeBorder: #3b82f6;
 $grayText: #9ca3af;
 $danger: #dc2626;
-$danger: #dc26268e;
+$dangerDark: #dc26268e;
 $gray: #6b7280;
+$grayDark: #4c515b;
 
-$font: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-sans-serif;
-
+$font: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
+  "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
 
 // 전체 레이아웃
 body {
@@ -47,6 +46,38 @@ body {
   box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.05);
 }
 
+// footer
+.footer {
+  position: fixed;
+  right: 0;
+  left: 0;
+  bottom: 0;
+  z-index: 10;
+  width: 100%;
+  max-width: 600px;
+  margin: 0 auto;
+  background-color: transparent;
+  color: $white;
+
+  &__grid {
+    display: grid;
+    grid-template-columns: repeat(3, minmax(0, 1fr));
+    background-color: transparent;
+
+    button {
+      border: none;
+      display: flex;
+      gap: 8px;
+      justify-content: center;
+      align-items: center;
+      font-size: 15px;
+      background-color: transparent;
+      padding: 16px 0px;
+      backdrop-filter: blur(10px);
+    }
+  }
+}
+
 // home 레이아웃
 .home {
   &__title {
@@ -54,23 +85,28 @@ body {
     font-weight: 700;
     font-size: 20px;
   }
+
   &__tabs {
     width: 100%;
     display: grid;
     grid-template-columns: repeat(2, minmax(0, 1fr));
   }
+
   &__tab {
     text-align: center;
     font-weight: 500;
     padding: 16px;
     cursor: pointer;
-  }
-  &--active {
-    font-weight: 700;
-    border-bottom: 2px solid $activeBorder;
-  }
-  &:hover, &:focus {
-    background-color: $activeBg;
+
+    &--active {
+      font-weight: 700;
+      border-bottom: 2px solid $activeBorder;
+    }
+
+    &:hover,
+    &:focus {
+      background-color: $activeBg;
+    }
   }
 }
 
@@ -85,7 +121,7 @@ body {
     resize: none;
     outline: none;
     font-size: 20px;
-    border-bottom: 1px solid &$border;
+    border-bottom: 1px solid $border;
 
     &::placeholder {
       font-family: $font;
@@ -117,5 +153,95 @@ body {
     @extend .button;
     padding: 0.5rem 1rem;
     background-color: $primaryColor;
+  }
+}
+
+// Post 레이아웃
+.post {
+  padding-bottom: 40px;
+
+  &__flex {
+    display: flex;
+    gap: 4px;
+    align-items: center;
+  }
+
+  &__email,
+  &__createdAt {
+    font-size: 12px;
+    color: $grayText;
+  }
+
+  &__box {
+    padding: 10px 0px;
+    border-bottom: 1px solid $border;
+
+    a {
+      text-decoration: none;
+    }
+
+    &-profile {
+      text-decoration: none;
+      padding: 0px 10px;
+
+      &-img,
+      &-icon {
+        width: 40px;
+        height: 40px;
+        border-radius: 50%;
+        margin-right: 10px;
+      }
+    }
+
+    &-content {
+      display: block;
+      font-weight: 500;
+      padding: 16px 4px;
+    }
+
+    &-footer {
+      padding: 0 16px;
+      margin-top: 10px;
+      font-size: 14px;
+      display: flex;
+      gap: 8px;
+      flex-direction: row-reverse;
+
+      button {
+        @extend .button;
+
+        &.post__delete {
+          background-color: $danger;
+
+          &:hover,
+          &:focus {
+            background-color: $dangerDark;
+          }
+        }
+
+        &.post__edit {
+          background-color: $gray;
+
+          a {
+            background-color: transparent;
+          }
+
+          &:hover,
+          &:focus {
+            background-color: $grayDark;
+          }
+        }
+
+        &.post__likes,
+        &.post__comments {
+          @extend .post__flex;
+
+          &:hover,
+          &:focus {
+            color: $gray;
+          }
+        }
+      }
+    }
   }
 }
